moved DataProps
